用户名登录

+++ admin-core/src/main/java/com/lb/service/IUserRoleService.java
@@ -4,11 +4,14 @@ import com.baomidou.mybatisplus.extension.service.IService;
 import com.lb.entity.UserRole;
 
 /**
- * 用户与角色对应关系(sys_user_role)表服务接口
+ * <p>
+ * 用户与角色对应关系 服务类
+ * </p>
  *
- * @author mybatis-generator
- * @since 2019-04-11 11:41:29
+ * @author null123
+ * @since 2019-04-11
  */
 public interface IUserRoleService extends IService<UserRole> {
-    UserRole getByUserId(Integer userId);
+
+    UserRole getByUserId(Long userId);
 }
